Create plan.dart
The Plan class will be used to reference a plan in the Firebase Firestore database. In turn, this reference will be responsible for pointing to a url of a file saved in Firebase Storage. In short, a Plan is an object with a defined type and url, while a plan is the file that will be opened in the program itself.